Clean-up, fixed bugs/inconsistency
Made the code more readable for users and easier to edit if necessary. Redone from scratch to eliminate unnecessary elements.

VIP page is bugged for now, not a page that is used so left it for now. All other things such as dashboard, search results, account, profile, settings etc all fixed.