Optional Style
I wasn't happy with the original styling of the website. As the code is somewhat finished, I'll leave it up and add this style as an additional option.

The colour palette has been changed slightly as-well as most of the VIP elements hidden without the need of adblock. Pages that are not used by myself such as discover, VIP, apps etc have been hidden for a cleaner look. If you use any of these, just remove the elements from the code. Change the elements that have the attributes display:none; or visibility: hidden.